# StringSegment: better performance