--- a/src/Primitives/src/ThrowHelper.cs
+++ b/src/Primitives/src/ThrowHelper.cs
@@ -89,6 +89,7 @@ namespace Microsoft.Extensions.Primitives
     internal enum ExceptionResource
     {
         Argument_InvalidOffsetLength,
+        Argument_InvalidOffsetLengthStringSegment,
         Capacity_CannotChangeAfterWriteStarted,
         Capacity_NotEnough,
         Capacity_NotUsedEntirely