# Description

* JIT produces better code
* enabled inlining on more methods
* mostly micro-optimizations

# Benchmarks

## Overview

``` ini

BenchmarkDotNet=v0.10.11, OS=ubuntu 17.10
Processor=Intel Xeon CPU 2.60GHz, ProcessorCount=2
.NET Core SDK=2.1.4
  [Host]     : .NET Core 2.0.5 (Framework 4.6.0.0), 64bit RyuJIT
  Job-DNRIOU : .NET Core 2.0.5 (Framework 4.6.0.0), 64bit RyuJIT

RemoveOutliers=False  Runtime=Core  Server=True  
Toolchain=.NET Core 2.0  LaunchCount=3  RunStrategy=Throughput  
TargetCount=10  WarmupCount=5  

```
|                Method |       Mean |     Error |    StdDev |            Op/s |  Gen 0 | Allocated |
|---------------------- |-----------:|----------:|----------:|----------------:|-------:|----------:|
|           Ctor_String |  6.1409 ns | 0.0188 ns | 0.0281 ns |   162,842,294.7 |      - |       0 B |
|              GetValue |  3.0951 ns | 0.0474 ns | 0.0709 ns |   323,092,110.5 |      - |       0 B |
|               Indexer |  0.2666 ns | 0.0134 ns | 0.0201 ns | 3,750,293,266.8 |      - |       0 B |
| Equals_Object_Invalid |  3.7666 ns | 0.3823 ns | 0.5722 ns |   265,489,382.6 |      - |       0 B |
|   Equals_Object_Valid | 11.6428 ns | 0.0364 ns | 0.0544 ns |    85,889,775.5 |      - |       0 B |
|          Equals_Valid |  5.6707 ns | 0.0348 ns | 0.0520 ns |   176,346,164.4 |      - |       0 B |
|         Equals_String |  5.7977 ns | 0.0231 ns | 0.0346 ns |   172,481,418.0 |      - |       0 B |
|           GetHashCode | 22.9720 ns | 0.1092 ns | 0.1634 ns |    43,531,342.3 |      - |       0 B |
|            StartsWith | 13.2762 ns | 0.0566 ns | 0.0847 ns |    75,322,957.9 |      - |       0 B |
|              EndsWith | 12.8829 ns | 0.0591 ns | 0.0885 ns |    77,622,457.0 |      - |       0 B |
|             SubString | 20.7859 ns | 0.1463 ns | 0.2190 ns |    48,109,488.1 | 0.0005 |      32 B |
|            SubSegment | 24.9102 ns | 0.0548 ns | 0.0820 ns |    40,144,255.7 |      - |       0 B |
|               IndexOf |  6.3147 ns | 0.0751 ns | 0.1124 ns |   158,359,446.9 |      - |       0 B |
|            IndexOfAny | 18.7092 ns | 0.1530 ns | 0.2290 ns |    53,449,551.3 |      - |       0 B |
|           LastIndexOf |  9.3210 ns | 0.0415 ns | 0.0622 ns |   107,285,050.6 |      - |       0 B |
|                  Trim | 78.5879 ns | 0.3100 ns | 0.4640 ns |    12,724,602.2 |      - |       0 B |
|             TrimStart | 43.0461 ns | 0.1081 ns | 0.1617 ns |    23,230,898.6 |      - |       0 B |
|               TrimEnd | 45.1374 ns | 0.1594 ns | 0.2385 ns |    22,154,559.9 |      - |       0 B |


## Details

[Code for benchmarks](https://github.com/gfoidl/Benchmarks/tree/e04edcb91c0bf2fea1a46c2afa8db9d321ca3773/aspnet/Common/StringSegment/StringSegment)

### Ctor (string)

``` ini

BenchmarkDotNet=v0.10.12, OS=Windows 10 Redstone 3 [1709, Fall Creators Update] (10.0.16299.248)
Intel Core i7-7700HQ CPU 2.80GHz (Kaby Lake), 1 CPU, 8 logical cores and 4 physical cores
Frequency=2742187 Hz, Resolution=364.6724 ns, Timer=TSC
.NET Core SDK=2.1.4
  [Host]     : .NET Core 2.0.5 (Framework 4.6.26020.03), 64bit RyuJIT
  DefaultJob : .NET Core 2.0.5 (Framework 4.6.26020.03), 64bit RyuJIT


```
|  Method |     Mean |     Error |    StdDev | Scaled | ScaledSD |
|-------- |---------:|----------:|----------:|-------:|---------:|
| Default | 2.013 ns | 0.0359 ns | 0.0318 ns |   1.00 |     0.00 |
|     New | 1.648 ns | 0.0280 ns | 0.0262 ns |   0.82 |     0.02 |

### Value

``` ini

BenchmarkDotNet=v0.10.12, OS=Windows 10 Redstone 3 [1709, Fall Creators Update] (10.0.16299.248)
Intel Core i7-7700HQ CPU 2.80GHz (Kaby Lake), 1 CPU, 8 logical cores and 4 physical cores
Frequency=2742187 Hz, Resolution=364.6724 ns, Timer=TSC
.NET Core SDK=2.1.4
  [Host]     : .NET Core 2.0.5 (Framework 4.6.26020.03), 64bit RyuJIT
  DefaultJob : .NET Core 2.0.5 (Framework 4.6.26020.03), 64bit RyuJIT


```
|  Method |     Mean |     Error |    StdDev | Scaled | ScaledSD |
|-------- |---------:|----------:|----------:|-------:|---------:|
| Default | 2.787 ns | 0.1408 ns | 0.1317 ns |   1.00 |     0.00 |
|     New | 2.560 ns | 0.0773 ns | 0.0685 ns |   0.92 |     0.05 |

### Equals (valid)

``` ini

BenchmarkDotNet=v0.10.12, OS=Windows 10 Redstone 3 [1709, Fall Creators Update] (10.0.16299.248)
Intel Core i7-7700HQ CPU 2.80GHz (Kaby Lake), 1 CPU, 8 logical cores and 4 physical cores
Frequency=2742187 Hz, Resolution=364.6724 ns, Timer=TSC
.NET Core SDK=2.1.4
  [Host]     : .NET Core 2.0.5 (Framework 4.6.26020.03), 64bit RyuJIT
  DefaultJob : .NET Core 2.0.5 (Framework 4.6.26020.03), 64bit RyuJIT


```
|  Method |     Mean |     Error |    StdDev | Scaled |
|-------- |---------:|----------:|----------:|-------:|
| Default | 4.935 ns | 0.0395 ns | 0.0350 ns |   1.00 |
|     New | 4.884 ns | 0.0298 ns | 0.0264 ns |   0.99 |

### Equals (string)

``` ini

BenchmarkDotNet=v0.10.12, OS=Windows 10 Redstone 3 [1709, Fall Creators Update] (10.0.16299.248)
Intel Core i7-7700HQ CPU 2.80GHz (Kaby Lake), 1 CPU, 8 logical cores and 4 physical cores
Frequency=2742187 Hz, Resolution=364.6724 ns, Timer=TSC
.NET Core SDK=2.1.4
  [Host]     : .NET Core 2.0.5 (Framework 4.6.26020.03), 64bit RyuJIT
  DefaultJob : .NET Core 2.0.5 (Framework 4.6.26020.03), 64bit RyuJIT


```
|  Method |     Mean |     Error |    StdDev | Scaled |
|-------- |---------:|----------:|----------:|-------:|
| Default | 7.097 ns | 0.0796 ns | 0.0744 ns |   1.00 |
|     New | 5.013 ns | 0.0540 ns | 0.0505 ns |   0.71 |

### GetHashCode

``` ini

BenchmarkDotNet=v0.10.12, OS=Windows 10 Redstone 3 [1709, Fall Creators Update] (10.0.16299.248)
Intel Core i7-7700HQ CPU 2.80GHz (Kaby Lake), 1 CPU, 8 logical cores and 4 physical cores
Frequency=2742187 Hz, Resolution=364.6724 ns, Timer=TSC
.NET Core SDK=2.1.4
  [Host]     : .NET Core 2.0.5 (Framework 4.6.26020.03), 64bit RyuJIT
  DefaultJob : .NET Core 2.0.5 (Framework 4.6.26020.03), 64bit RyuJIT


```
|  Method |     Mean |     Error |    StdDev | Scaled | ScaledSD | Allocated |
|-------- |---------:|----------:|----------:|-------:|---------:|----------:|
| Default | 16.46 ns | 0.3662 ns | 0.3919 ns |   1.00 |     0.00 |       0 B |
|     New | 16.18 ns | 0.3075 ns | 0.2876 ns |   0.98 |     0.03 |       0 B |

### StartsWith

``` ini

BenchmarkDotNet=v0.10.12, OS=Windows 10 Redstone 3 [1709, Fall Creators Update] (10.0.16299.248)
Intel Core i7-7700HQ CPU 2.80GHz (Kaby Lake), 1 CPU, 8 logical cores and 4 physical cores
Frequency=2742187 Hz, Resolution=364.6724 ns, Timer=TSC
.NET Core SDK=2.1.4
  [Host]     : .NET Core 2.0.5 (Framework 4.6.26020.03), 64bit RyuJIT
  DefaultJob : .NET Core 2.0.5 (Framework 4.6.26020.03), 64bit RyuJIT


```
|  Method |     Mean |     Error |    StdDev | Scaled |
|-------- |---------:|----------:|----------:|-------:|
| Default | 14.18 ns | 0.2111 ns | 0.1974 ns |   1.00 |
|     New | 12.31 ns | 0.0919 ns | 0.0860 ns |   0.87 |

### EndsWith

``` ini

BenchmarkDotNet=v0.10.12, OS=Windows 10 Redstone 3 [1709, Fall Creators Update] (10.0.16299.248)
Intel Core i7-7700HQ CPU 2.80GHz (Kaby Lake), 1 CPU, 8 logical cores and 4 physical cores
Frequency=2742187 Hz, Resolution=364.6724 ns, Timer=TSC
.NET Core SDK=2.1.4
  [Host]     : .NET Core 2.0.5 (Framework 4.6.26020.03), 64bit RyuJIT
  DefaultJob : .NET Core 2.0.5 (Framework 4.6.26020.03), 64bit RyuJIT


```
|  Method |     Mean |     Error |    StdDev | Scaled |
|-------- |---------:|----------:|----------:|-------:|
| Default | 13.42 ns | 0.1648 ns | 0.1542 ns |   1.00 |
|     New | 11.56 ns | 0.1644 ns | 0.1538 ns |   0.86 |

### Substring

``` ini

BenchmarkDotNet=v0.10.12, OS=Windows 10 Redstone 3 [1709, Fall Creators Update] (10.0.16299.248)
Intel Core i7-7700HQ CPU 2.80GHz (Kaby Lake), 1 CPU, 8 logical cores and 4 physical cores
Frequency=2742187 Hz, Resolution=364.6724 ns, Timer=TSC
.NET Core SDK=2.1.4
  [Host]     : .NET Core 2.0.5 (Framework 4.6.26020.03), 64bit RyuJIT
  DefaultJob : .NET Core 2.0.5 (Framework 4.6.26020.03), 64bit RyuJIT


```
|  Method |     Mean |     Error |    StdDev | Scaled | ScaledSD |
|-------- |---------:|----------:|----------:|-------:|---------:|
| Default | 13.91 ns | 0.3772 ns | 0.7532 ns |   1.00 |     0.00 |
|     New | 11.53 ns | 0.3217 ns | 0.4716 ns |   0.83 |     0.06 |

### Subsegment

``` ini

BenchmarkDotNet=v0.10.12, OS=Windows 10 Redstone 3 [1709, Fall Creators Update] (10.0.16299.248)
Intel Core i7-7700HQ CPU 2.80GHz (Kaby Lake), 1 CPU, 8 logical cores and 4 physical cores
Frequency=2742187 Hz, Resolution=364.6724 ns, Timer=TSC
.NET Core SDK=2.1.4
  [Host]     : .NET Core 2.0.5 (Framework 4.6.26020.03), 64bit RyuJIT
  DefaultJob : .NET Core 2.0.5 (Framework 4.6.26020.03), 64bit RyuJIT


```
|  Method |     Mean |     Error |    StdDev | Scaled | ScaledSD |
|-------- |---------:|----------:|----------:|-------:|---------:|
| Default | 4.865 ns | 0.1125 ns | 0.1053 ns |   1.00 |     0.00 |
|     New | 4.642 ns | 0.0430 ns | 0.0381 ns |   0.95 |     0.02 |

### IndexOf

``` ini

BenchmarkDotNet=v0.10.12, OS=Windows 10 Redstone 3 [1709, Fall Creators Update] (10.0.16299.248)
Intel Core i7-7700HQ CPU 2.80GHz (Kaby Lake), 1 CPU, 8 logical cores and 4 physical cores
Frequency=2742187 Hz, Resolution=364.6724 ns, Timer=TSC
.NET Core SDK=2.1.4
  [Host]     : .NET Core 2.0.5 (Framework 4.6.26020.03), 64bit RyuJIT
  DefaultJob : .NET Core 2.0.5 (Framework 4.6.26020.03), 64bit RyuJIT


```
|  Method |     Mean |     Error |    StdDev | Scaled | ScaledSD |
|-------- |---------:|----------:|----------:|-------:|---------:|
| Default | 15.23 ns | 0.3410 ns | 0.4188 ns |   1.00 |     0.00 |
|     New | 13.14 ns | 0.2340 ns | 0.2074 ns |   0.86 |     0.03 |

### IndexOfAny

``` ini

BenchmarkDotNet=v0.10.12, OS=Windows 10 Redstone 3 [1709, Fall Creators Update] (10.0.16299.248)
Intel Core i7-7700HQ CPU 2.80GHz (Kaby Lake), 1 CPU, 8 logical cores and 4 physical cores
Frequency=2742187 Hz, Resolution=364.6724 ns, Timer=TSC
.NET Core SDK=2.1.4
  [Host]     : .NET Core 2.0.5 (Framework 4.6.26020.03), 64bit RyuJIT
  DefaultJob : .NET Core 2.0.5 (Framework 4.6.26020.03), 64bit RyuJIT


```
|  Method |     Mean |     Error |    StdDev | Scaled |
|-------- |---------:|----------:|----------:|-------:|
| Default | 35.26 ns | 0.2703 ns | 0.2528 ns |   1.00 |
|     New | 34.36 ns | 0.2044 ns | 0.1812 ns |   0.97 |

### LastIndexOf

``` ini

BenchmarkDotNet=v0.10.12, OS=Windows 10 Redstone 3 [1709, Fall Creators Update] (10.0.16299.248)
Intel Core i7-7700HQ CPU 2.80GHz (Kaby Lake), 1 CPU, 8 logical cores and 4 physical cores
Frequency=2742187 Hz, Resolution=364.6724 ns, Timer=TSC
.NET Core SDK=2.1.4
  [Host]     : .NET Core 2.0.5 (Framework 4.6.26020.03), 64bit RyuJIT
  DefaultJob : .NET Core 2.0.5 (Framework 4.6.26020.03), 64bit RyuJIT


```
|  Method |     Mean |     Error |    StdDev | Scaled |
|-------- |---------:|----------:|----------:|-------:|
| Default | 15.58 ns | 0.1433 ns | 0.1341 ns |   1.00 |
|     New | 14.98 ns | 0.1969 ns | 0.1842 ns |   0.96 |

### TrimStart

``` ini

BenchmarkDotNet=v0.10.12, OS=Windows 10 Redstone 3 [1709, Fall Creators Update] (10.0.16299.248)
Intel Core i7-7700HQ CPU 2.80GHz (Kaby Lake), 1 CPU, 8 logical cores and 4 physical cores
Frequency=2742187 Hz, Resolution=364.6724 ns, Timer=TSC
.NET Core SDK=2.1.4
  [Host]     : .NET Core 2.0.5 (Framework 4.6.26020.03), 64bit RyuJIT
  DefaultJob : .NET Core 2.0.5 (Framework 4.6.26020.03), 64bit RyuJIT


```
|  Method |     Mean |     Error |    StdDev | Scaled |
|-------- |---------:|----------:|----------:|-------:|
| Default | 80.59 ns | 0.7632 ns | 0.6373 ns |   1.00 |
|     New | 69.57 ns | 0.6849 ns | 0.6407 ns |   0.86 |

### TrimEnd

``` ini

BenchmarkDotNet=v0.10.12, OS=Windows 10 Redstone 3 [1709, Fall Creators Update] (10.0.16299.248)
Intel Core i7-7700HQ CPU 2.80GHz (Kaby Lake), 1 CPU, 8 logical cores and 4 physical cores
Frequency=2742187 Hz, Resolution=364.6724 ns, Timer=TSC
.NET Core SDK=2.1.4
  [Host]     : .NET Core 2.0.5 (Framework 4.6.26020.03), 64bit RyuJIT
  DefaultJob : .NET Core 2.0.5 (Framework 4.6.26020.03), 64bit RyuJIT


```
|  Method |     Mean |     Error |    StdDev | Scaled | ScaledSD |
|-------- |---------:|----------:|----------:|-------:|---------:|
| Default | 50.26 ns | 1.0791 ns | 1.0094 ns |   1.00 |     0.00 |
|     New | 43.75 ns | 0.5553 ns | 0.5195 ns |   0.87 |     0.02 |

# Discussion

## Span-based implementation

I tried a [Span-based version](https://github.com/gfoidl/Benchmarks/blob/disc/stringsegment-span/aspnet/Common/StringSegment/StringSegment/StringSegments/StringSegment2.cs) for `IndexOf`, but this doubled the time. So this approach was discontinued in this PR.