@davidfowl 2018-03-14
@gfoidl can you rebase this?