@davidfowl 2018-03-24
> I tried a Span-based version for IndexOf, but this doubled the time. So this approach was discontinued in this PR.

That's interesting since IndexOf is vectorized. Any ideas @ahsonkhan @benaadams @AndyAyersMS