@gfoidl 2018-03-25
@AndyAyersMS the code is in https://github.com/gfoidl/Benchmarks/blob/504fd553530e5399b218822773fec9ce78932d6c/aspnet/Common/StringSegment/StringSegment/StringSegments/StringSegment2.cs#L426

When I ran a trivial benchmark
```c#
public class Benchmarks
{
    private string _input;

    public Benchmarks()
    {
        const int n = 100;
        _input = string.Create(n, 'a', (span, state) =>
        {
            span.Fill(state);
            span[n - 1] = 'c';
        });
    }

    [Benchmark(Baseline = true)]
    public int String_IndexOf() => _input.IndexOf('c');

    [Benchmark]
    public int Span_IndexOf() => _input.AsSpan().IndexOf('c');
}
```
the results are similar:
``` ini

BenchmarkDotNet=v0.10.13, OS=Windows 10 Redstone 3 [1709, Fall Creators Update] (10.0.16299.309)
Intel Core i7-7700HQ CPU 2.80GHz (Kaby Lake), 1 CPU, 8 logical cores and 4 physical cores
Frequency=2742189 Hz, Resolution=364.6722 ns, Timer=TSC
.NET Core SDK=2.1.300-preview3-008384
  [Host]     : .NET Core 2.1.0-preview2-26313-01 (CoreCLR 4.6.26310.01, CoreFX 4.6.26313.01), 64bit RyuJIT
  DefaultJob : .NET Core 2.1.0-preview2-26313-01 (CoreCLR 4.6.26310.01, CoreFX 4.6.26313.01), 64bit RyuJIT


```
|         Method |     Mean |     Error |    StdDev | Scaled | ScaledSD |
|--------------- |---------:|----------:|----------:|-------:|---------:|
| String_IndexOf | 17.06 ns | 0.3716 ns | 0.4961 ns |   1.00 |     0.00 |
|   Span_IndexOf | 38.00 ns | 0.7882 ns | 0.7373 ns |   2.23 |     0.07 |

The larger `n` is in the benchmark, the worse the results for the Span-version. 

`string.IndexOf` is also vectorized?
I believe a assumption that this is due any overhead of the span-based approach won't hold, because then it should not scale with the search-length.