@AndyAyersMS 2018-03-25
Yes, somthething odd is going on with the span version. I'll take a deeper look when I havea chance.

@ahsonkhan any ideas?