@ahsonkhan 2018-03-25
> That's interesting since IndexOf is vectorized. 
> Yes, somthething odd is going on with the span version.

Yes, the Span IndexOf is only vectorized for T = byte atm. For T != byte, we end up calling the generic method which is only loop unrolled and calls the .Equals method): https://github.com/dotnet/corefx/blob/master/src/Common/src/CoreLib/System/SpanHelpers.T.cs#L53

We have an issue open to port back the optimization for T = char, and possibly other primitive types (among other optimizations): https://github.com/dotnet/corefx/issues/25893

Also, we recently vectorized string.IndexOf by porting the Span (byte) optimization over:
http://source.dot.net/#System.Private.CoreLib/shared/System/String.Searching.cs,75 

That is the reason there is a performance difference, specifically for string.AsSpan().IndexOf, since string.AsSpan() returns a `ReadOnlySpan<char>`