@ahsonkhan 2018-03-25
@gfoidl, out of curiosity, can you please re-run your last benchmark and share the results of the comparison, but instead of using Span.IndexOf, can you use this method instead overload (wrap the single character into a ROS, and pass in StringComparison.Ordinal):
```C#
public static int IndexOf(this ReadOnlySpan<char> span, ReadOnlySpan<char> value, StringComparison comparisonType);
```