@gfoidl 2018-03-25
@ahsonkhan  you mean
```c#
public class Benchmarks
{
    private string _input;
    private char[] _value;

    public Benchmarks()
    {
        const int n = 100;
        _input = string.Create(n, 'a', (span, state) =>
        {
            span.Fill(state);
            span[n - 1] = 'c';
        });

        _value = new[] { 'c' };
    }

    [Benchmark(Baseline = true)]
    public int String_IndexOf() => _input.IndexOf('c');

    [Benchmark]
    public int Span_IndexOf() => _input.AsSpan().IndexOf('c');

    [Benchmark]
    public int Span_IndexOf_Ordinal() => _input.AsSpan().IndexOf(_value, StringComparison.Ordinal);
}
```

This results in

``` ini

BenchmarkDotNet=v0.10.13, OS=Windows 10 Redstone 3 [1709, Fall Creators Update] (10.0.16299.309)
Intel Core i7-7700HQ CPU 2.80GHz (Kaby Lake), 1 CPU, 8 logical cores and 4 physical cores
Frequency=2742189 Hz, Resolution=364.6722 ns, Timer=TSC
.NET Core SDK=2.1.300-preview3-008384
  [Host]     : .NET Core 2.1.0-preview2-26313-01 (CoreCLR 4.6.26310.01, CoreFX 4.6.26313.01), 64bit RyuJIT
  DefaultJob : .NET Core 2.1.0-preview2-26313-01 (CoreCLR 4.6.26310.01, CoreFX 4.6.26313.01), 64bit RyuJIT


```
|               Method |      Mean |     Error |    StdDev | Scaled | ScaledSD |
|--------------------- |----------:|----------:|----------:|-------:|---------:|
|       String_IndexOf |  13.86 ns | 0.0992 ns | 0.0880 ns |   1.00 |     0.00 |
|         Span_IndexOf |  37.30 ns | 0.2523 ns | 0.2237 ns |   2.69 |     0.02 |
| Span_IndexOf_Ordinal | 187.01 ns | 2.3025 ns | 2.0411 ns |  13.49 |     0.16 |

(I ran it twice, to prove that the numbers hold).