@ahsonkhan 2018-03-25
> you mean

Yes, thank you. It just verifies what I had expected. We can, in addition, optimize the StringComparison.Ordinal (along with OrdinalIgnoreCase) cases for IndexOf instead of calling down to the interop layer.