@gfoidl 2018-03-26
When I try the code from https://github.com/dotnet/corefx/pull/28464 (great -- thanks), I'll get

``` ini

BenchmarkDotNet=v0.10.13, OS=Windows 10 Redstone 3 [1709, Fall Creators Update] (10.0.16299.309)
Intel Core i7-7700HQ CPU 2.80GHz (Kaby Lake), 1 CPU, 8 logical cores and 4 physical cores
Frequency=2742189 Hz, Resolution=364.6722 ns, Timer=TSC
.NET Core SDK=2.1.300-preview3-008384
  [Host]     : .NET Core 2.1.0-preview2-26313-01 (CoreCLR 4.6.26310.01, CoreFX 4.6.26313.01), 64bit RyuJIT
  DefaultJob : .NET Core 2.1.0-preview2-26313-01 (CoreCLR 4.6.26310.01, CoreFX 4.6.26313.01), 64bit RyuJIT


```
|                  Method |     Mean |     Error |    StdDev | Scaled | ScaledSD |
|------------------------ |---------:|----------:|----------:|-------:|---------:|
|          String_IndexOf | 16.77 ns | 0.1015 ns | 0.0899 ns |   1.00 |     0.00 |
|            Span_IndexOf | 38.61 ns | 1.0304 ns | 1.0120 ns |   2.30 |     0.06 |
| Span_IndexOf_Vectorized | 20.34 ns | 0.5462 ns | 0.5109 ns |   1.21 |     0.03 |

```c#
[Benchmark]
public int Span_IndexOf_Vectorized()
{
    ref char searchSpace = ref MemoryMarshal.GetReference(_input.AsSpan());
    return IndexOf(ref searchSpace, 'c', _input.Length);
}
```

Note: I copy & pasted the `IndexOf`-method to the benchmark-app. Your change isn't available in the daily builds and on building from source I can't see this change neither (maybe I did something wrong there).