@ahsonkhan 2018-03-28
Thanks @gfoidl! This PR may be of interest to you: https://github.com/aspnet/Common/pull/324

The objective (and result) here is that string/span.IndexOf should have almost identical perf characteristics.