@gfoidl 2018-04-04
With https://github.com/dotnet/coreclr/pull/17284 `string.IndexOf` calls `SpanHelpers.IndexOf`. 
Is there any point in "user-code" (like this one) to use Span over string? 

There shouldn't be any difference in perf -- it's the same code working, the overhead of the Span-version should be negligible (`AsSpan` vs. `ref _firstChar`).

My question in essence is: should I update this PR to use `Span.IndexOf` or stick by `string.IndexOf`?