@ahsonkhan 2018-04-04
> Is there any point in "user-code" (like this one) to use Span over string?

It depends, but in general, if you use spans across the board, you will avoid the allocations. If, however, you have to end up getting a string back anyway, and IndexOf is the only span API you are using (and not something like Slice), then string should be OK to use.